Move discovery code out of DrupalKernelTrait and into LegacyServiceInstantiator.
